Add tests
This adds nearly complete test coverage. In addition, the redeploy
command has been changed to redeploy the entire cluster instead of just
a single service. This is necessary to avoid any sort of race
conditions.

I have also added tox config. It isn't necessary to use tox for local
development, but I find it easier to get travis to work with tox.