Support worker autoscaling
Adding autoscaling to ECS for the worker service requires a few changes
to how the cluster is redeployed. Previously, everything was stopped by
setting the desiredCount to 0 and then started back up by setting the
desiredCount to 1. With ECS controlling the scaling for the worker
service it's no longer possible to rely on the desiredCount of the
worker service remaining at 0 until things are ready to be started
again.

The change here sets the desiredCount of the scheduler to 0 first, then
does a --force-new-deployment for the web and worker services, then
finally does a --force-new-deployment for the scheduler. While
redeploying becomes more complicated, there's no need anymore to track
the current worker count. We can just set it to 1 and let ECS scale as
needed.

I've also added stop-scheduler and start-scheduler commands which can be
used in conjunction with deploying a new service definition with
Terraform.