Add Github Actions
This is an initial attempt at setting up Github Actions. It's not
optimal, but it does the job until I can figure out better ways to do
things. The biggest issue is that I think the tests workflow should
probably be turned into an action since it's reused in both the test and
deploy workflows.

On push the usual CI jobs will run. On merging into master the CI jobs
will run and then, if successful, a new image for the workflow container
will be built and deployed. The staging Airflow cluster will be
redeployed with the new image.