Remove travis config
Github Actions seems to be working fine. The only downside so far is
that the coveralls integration is minimal. It will correctly send
coverage info to coveralls, but the build won't fail if the coverage has
decreased. I don't think this is really a blocker and it seems likely
the integration will improve as Actions sees more use.