Add GeoWeb MARC data load
This adds a monthly workflow to load Aleph records with spatial data
into GeoWeb. It replaces an old process that is no longer possible due
to the migration to AWS.

The records are loaded on the 2nd of every month and pulled from the
same full MARC export to S3 that mario is using.